+++ CHANGELOG.md
@@ -27,3 +27,4 @@ v0.0.1
 * Added (optional) replacement for "Add block" button, showing blocks in a categorized popup
 * Fixed PHP-CS-Fixer config
 * Added integration of a liipMonitor dashboard
+* Add overriding of lexik maintenance to lock and unlock by site